pyproject.toml
[project]
name = "server-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "httpx>=0.28.1",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
